+++ features/support/world.ts
@@ -10,6 +10,7 @@ import VError from 'verror'
 import _ from 'lodash'
 import ndjsonParse from 'ndjson-parse'
 import { messages } from '@cucumber/messages'
+import FakeReportServer from '../../test/fake_report_server'
 
 interface ILastRun {
   error: any
@@ -32,8 +33,13 @@ export class World {
   public verifiedLastRunError: boolean
   public localExecutablePath: string
   public globalExecutablePath: string
+  public reportServer: FakeReportServer
 
-  async run(executablePath: string, inputArgs: string[]): Promise<void> {
+  async run(
+    executablePath: string,
+    inputArgs: string[],
+    env: NodeJS.ProcessEnv = process.env
+  ): Promise<void> {
     const messageFilename = 'message.ndjson'
     const args = ['node', executablePath]
       .concat(inputArgs, [
@@ -54,9 +60,14 @@ export class World {
 
     if (this.spawn) {
       result = await new Promise((resolve) => {
-        execFile(args[0], args.slice(1), { cwd }, (error, stdout, stderr) => {
-          resolve({ error, stdout, stderr })
-        })
+        execFile(
+          args[0],
+          args.slice(1),
+          { cwd, env },
+          (error, stdout, stderr) => {
+            resolve({ error, stdout, stderr })
+          }
+        )
       })
     } else {
       const stdout = new PassThrough()