+++ src/cli/configuration_builder.ts
@@ -11,6 +11,7 @@ import glob from 'glob'
 import { promisify } from 'util'
 import { IPickleFilterOptions } from '../pickle_filter'
 import { IRuntimeOptions } from '../runtime'
+import { valueOrDefault } from '../value_checker'
 
 export interface IConfigurationFormat {
   outputTo: string
@@ -22,6 +23,7 @@ export interface IConfiguration {
   featurePaths: string[]
   formats: IConfigurationFormat[]
   formatOptions: IParsedArgvFormatOptions
+  publishing: boolean
   listI18nKeywordsFor: string
   listI18nLanguages: boolean
   order: string
@@ -33,12 +35,15 @@ export interface IConfiguration {
   shouldExitImmediately: boolean
   supportCodePaths: string[]
   supportCodeRequiredModules: string[]
+  suppressPublishAdvertisement: boolean
 }
 
 export interface INewConfigurationBuilderOptions {
   argv: string[]
   cwd: string
 }
+
+const DEFAULT_CUCUMBER_PUBLISH_URL = 'https://messages.cucumber.io/api/reports'
 
 export default class ConfigurationBuilder {
   static async build(
@@ -83,6 +88,7 @@ export default class ConfigurationBuilder {
       featurePaths,
       formats: this.getFormats(),
       formatOptions: this.options.formatOptions,
+      publishing: this.isPublishing(),
       listI18nKeywordsFor,
       listI18nLanguages,
       order: this.options.order,
@@ -108,6 +114,7 @@ export default class ConfigurationBuilder {
       shouldExitImmediately: this.options.exit,
       supportCodePaths,
       supportCodeRequiredModules: this.options.requireModule,
+      suppressPublishAdvertisement: this.isPublishAdvertisementSuppressed(),
     }
   }
 
@@ -157,13 +164,43 @@ export default class ConfigurationBuilder {
     return _.uniq(featureDirs)
   }
 
+  isPublishing(): boolean {
+    return (
+      this.options.publish ||
+      this.isTruthyString(process.env.CUCUMBER_PUBLISH_ENABLED) ||
+      process.env.CUCUMBER_PUBLISH_TOKEN !== undefined
+    )
+  }
+
+  isPublishAdvertisementSuppressed(): boolean {
+    return (
+      this.options.publishQuiet ||
+      this.isTruthyString(process.env.CUCUMBER_PUBLISH_QUIET)
+    )
+  }
+
   getFormats(): IConfigurationFormat[] {
     const mapping: { [key: string]: string } = { '': 'progress' }
     this.options.format.forEach((format) => {
       const [type, outputTo] = OptionSplitter.split(format)
       mapping[outputTo] = type
     })
+    if (this.isPublishing()) {
+      const publishUrl = valueOrDefault(
+        process.env.CUCUMBER_PUBLISH_URL,
+        DEFAULT_CUCUMBER_PUBLISH_URL
+      )
+
+      mapping[publishUrl] = 'message'
+    }
     return _.map(mapping, (type, outputTo) => ({ outputTo, type }))
+  }
+
+  isTruthyString(s: string | undefined): boolean {
+    if (s === undefined) {
+      return false
+    }
+    return s.match(/^(false|no|0)$/i) === null
   }
 
   async getUnexpandedFeaturePaths(): Promise<string[]> {