+++ src/support_code_library_builder/types.ts
@@ -13,10 +13,10 @@ export interface ITestCaseHookParameter {
   testCaseStartedId: string
 }
 
-export type TestCaseHookFunctionWithoutParameter = () => void
+export type TestCaseHookFunctionWithoutParameter = () => void | Promise<void>
 export type TestCaseHookFunctionWithParameter = (
   arg: ITestCaseHookParameter
-) => void
+) => void | Promise<void>
 export type TestCaseHookFunction =
   | TestCaseHookFunctionWithoutParameter
   | TestCaseHookFunctionWithParameter